+++ site/dofile.js
@@ -1,7 +1,0 @@
-// cacheless require
-
-module.exports = function(mod) {
-	delete require.cache[require.resolve(mod)];
-	return require(mod);
-};
-